Delete the verification of users who are no longer on the server
This mainly to the new addition of functionality when the bot gets back online, which avoids the necessity of the user to verify if any user on the database is no longer on the server by using the ~list command. Now the bot does this by itself, so this verification is no longer needed.