Complete the functionality of 'OnAbsence' while the bot was absent
Added the ability to track which servers are not available anymore after reconnecting to the Discord gateway. This makes possible to keep track of not available servers while the bot was offline, but due to this addition, the logic behind the tracking of the new servers and missing users while offline had to be merged on the same event with this new enhancement of the bot. Also, an overload of the ExecuteQuery function was created due to the new table added to the directory to keep track of the servers and had to accept strings instead of ulongs.