Fix prefix check not accepting the default bot prefix and @bot prefix
The bot was only accepting commands if the prefix was the one set on the server database, but in case that some user doesn't know what was the prefix, the bot should work with the default prefix and with @mention too, but it wasn't taking the last two into account. Deleted the file exists verification, since the bot handles all the files issues when connecting, and with the previous said, established the bot to accept the database prefix, the default prefix and the @mention prefix.